requirements.txt
flask==3.0.2
tensorflow==2.16.1
numpy==1.26.4
pillow==10.2.0
split-folders==0.5.1
opencv-python-headless==4.9.0.80
setuptools==69.1.0